Update app.js
Addition of MD5 check for the SecNodeTracker.js to prevent user modification of the file in an attempt to circumvent downtime, response times ect.

This value would need to be passed to the tracking servers for confirmation of the current MD5 value.  If it does not match, it can be assumed that the user has modified the SecNodeTracker.js application.

An exception would be generated while this is incorrect.