Add localaddress setting to bind to a network address
Node's net package allows to pass a localAddress option that binds the
socket connection to an specific address of the network interface.
Without it, the OS choose one (usually the firs address of the
interface).

This commit reads this option `localaddress` from the local storage
settings (even tough it does not ask for it in the setup) and passes
it into all the invocations that potentially opens a connection.

In order for this to work the connection to the server was forced
to always use the websocket transport, because otherwise the socket.io
client will first try using the polling before upgrading to a websocket
connection transport, which internally uses a XMLHttpRequest for the
polling and the socket-io options never reach that "deep" in the
module dependencies.